Split test_sending_messages into two modules
Split the module into 'test_available_modules' and 'test_messages'
